refactor: 웹 관련 코드를 web 폴더로 이동
